+++ desktop/src/main/java/bisq/desktop/primary/main/content/social/exchange/ExchangeModel.java
@@ -24,7 +24,4 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Getter
 public class ExchangeModel extends ChatModel {
-    public ExchangeModel() {
-        super();
-    }
 }
